fix circular dependencies in AddressForm component
